Refactoring, added KNN algorithms as submodules
